Redesign dal

+++ Common/Common.SearchParams/Core/SearchResult.cs
@@ -2,14 +2,14 @@
 
 public sealed class SearchResult<T>
 {
-    public int Total { get; set; }
+    public long Total { get; set; }
     public IList<T> Objects { get; set; } = [];
     public int RequestedPage { get; set; }
     public int? RequestedObjectsCount { get; set; }
 
     public SearchResult() { }
 
-    public SearchResult(int total, IEnumerable<T> objects, int requestedPage, int? requestedObjectsCount)
+    public SearchResult(long total, IEnumerable<T> objects, int requestedPage, int? requestedObjectsCount)
     {
         Total = total;
         Objects = [.. objects];
